This is where I (Cameron) will do my changes.
I have started by bringing in the model necessary to predict whether or not a charge is fraudulent.

I have made some changes and notes in the Home Controller denoting where I am going with my chagnes

my priority right now is building the functionality to predict whether or not a transaction is fraudulent or not.